Fire-Build: Script: Import A70Q Custom Script

+++ fire.sh
@@ -2,31 +2,75 @@
 
 set -e
 
-# Put the Kernel Path in here and fire it up
-# No need to change the whole and entire file to set-up kernel path
-KERNEL_PATH=/home/mydadisbrave2/m51kernel
-TOOLCHAIN_PATH=/home/mydadisbrave2
+echo "FireKernel Builder"
 
-# Export GCC and ARCH
+#
+## Put the Kernel Path in here and fire it up
+## No need to change the whole and entire file to set-up kernel path
+#
+KERNEL_PATH=$(pwd)
+TOOLCHAIN_PATH=/home/$USER
+BINARIES_OUT_PATH=out/arch/arm64/boot
+COMPLETE_OUT_PATH=$KERNEL_PATH/$BINARIES_OUT_PATH
+DEFCONFIG=m51_eur_open_defconfig
+
+#
+##For separated GCC & Clang Path
+#
+# GCC_PATH=
+# CLANG_PATH=
+
+#
+## Export GCC and ARCH
+#
 export CROSS_COMPILE=$TOOLCHAIN_PATH/aarch64-linux-android-4.9/bin/aarch64-linux-android-
 export ARCH=arm64
 
-# CleanUP Output Folder
-rm -rf out
+#
+## For dtbo & dtbs compile
+#
+if [ "$2" == "do-overlay" ]; then
+	export CONFIG_BUILD_ARM64_DT_OVERLAY=y
+fi
 
-# Output hacking & tricking
+#
+## Clean OutPut Folder
+#
+if [ "$1" == "do-clean" ]; then
+	rm -rf out
+fi
+
+#
+## Clean UP anykernel3 old output binaries & flash zips
+#
+if [ "$1" == "do-clean" ]; then
+        rm -rf AnyKernel3/*.zip AnyKernel3/*.gz-dtb AnyKernel3/dtbo.img
+fi
+
+#
+## Output hacking & tricking
+#
 if [ ! -d out ]; then
 	mkdir out
 fi
 
-# Types, paths, and more etc.
+#
+## Types, paths, and more etc.
+#
 BUILD_CROSS_COMPILE=$TOOLCHAIN_PATH/aarch64-linux-android-4.9/bin/aarch64-linux-android-
-KERNEL_LLVM_BIN=$TOOLCHAIN_PATH/clang-r353983c/bin/clang
+KERNEL_LLVM_BIN=$TOOLCHAIN_PATH/clang-r377782d/bin/clang
 CLANG_TRIPLE=aarch64-linux-gnu-
-KERNEL_MAKE_ENV="CONFIG_BUILD_ARM64_DT_OVERLAY=y"
 
-make -C $KERNEL_PATH O=$KERNEL_PATH/out ARCH=arm64 CROSS_COMPILE=$BUILD_CROSS_COMPILE REAL_CC=$KERNEL_LLVM_BIN CLANG_TRIPLE=$CLANG_TRIPLE m51_eur_open_defconfig
-make -j8 -C $KERNEL_PATH O=$KERNEL_PATH/out ARCH=arm64 CROSS_COMPILE=$BUILD_CROSS_COMPILE REAL_CC=$KERNEL_LLVM_BIN CLANG_TRIPLE=$CLANG_TRIPLE
+make -C $(pwd) O=$(pwd)/out ARCH=arm64 CROSS_COMPILE=$BUILD_CROSS_COMPILE CC=$KERNEL_LLVM_BIN CLANG_TRIPLE=$CLANG_TRIPLE $DEFCONFIG
+make -j8 -C $(pwd) O=$(pwd)/out ARCH=arm64 CROSS_COMPILE=$BUILD_CROSS_COMPILE CC=$KERNEL_LLVM_BIN CLANG_TRIPLE=$CLANG_TRIPLE
 
-tools/mkdtimg create out/arch/arm64/boot/dtbo.img --page_size=4096 $(find out -name "*.dtbo")
-tools/mkdtimg create out/arch/arm64/boot/recovery_dtbo --page_size=4096 $(find out -name "*.dtbo")
+if [ "$2" == "do-overlay" ]; then
+	tools/mkdtimg create $BINARIES_OUT_PATH/dtbo.img --page_size=4096 $(find out -name "*.dtbo")
+fi
+
+#
+## Copy Image.gz-dtb into anykernel3 folder [WIP]
+#
+#fi [ "$3" == "do-copy" ]; then
+#	cp $COMPLETE_OUT_PATH/Image.gz-dtb AnyKernel3/Image.gz-dtb
+#if
